Move settings classes to packages
