+++ app/src/main/java/org/mozilla/fenix/settings/deletebrowsingdata/DeleteAndQuit.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.utils
+package org.mozilla.fenix.settings.deletebrowsingdata
 
 import android.app.Activity
 import kotlinx.coroutines.CoroutineScope
@@ -11,9 +11,6 @@ import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.mozilla.fenix.ext.settings
-import org.mozilla.fenix.settings.DefaultDeleteBrowsingDataController
-import org.mozilla.fenix.settings.DeleteBrowsingDataController
-import org.mozilla.fenix.settings.DeleteBrowsingDataOnQuitType
 
 /**
  * Deletes selected browsing data and finishes the activity.
