+++ app/src/main/java/org/mozilla/fenix/settings/deletebrowsingdata/DeleteBrowsingDataItem.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.settings
+package org.mozilla.fenix.settings.deletebrowsingdata
 
 import android.content.Context
 import android.util.AttributeSet
