+++ app/src/main/java/org/mozilla/fenix/settings/sitepermissions/SitePermissionsFragment.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.settings
+package org.mozilla.fenix.settings.sitepermissions
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
@@ -13,6 +13,7 @@ import androidx.preference.PreferenceFragmentCompat
 import org.mozilla.fenix.R
 import org.mozilla.fenix.ext.getPreferenceKey
 import org.mozilla.fenix.ext.settings
+import org.mozilla.fenix.settings.PhoneFeature
 
 @SuppressWarnings("TooManyFunctions")
 class SitePermissionsFragment : PreferenceFragmentCompat() {
@@ -69,7 +70,8 @@ class SitePermissionsFragment : PreferenceFragmentCompat() {
     }
 
     private fun navigateToPhoneFeature(phoneFeature: PhoneFeature) {
-        val directions = SitePermissionsFragmentDirections.actionSitePermissionsToManagePhoneFeatures(phoneFeature.id)
+        val directions = SitePermissionsFragmentDirections
+            .actionSitePermissionsToManagePhoneFeatures(phoneFeature.id)
         Navigation.findNavController(view!!).navigate(directions)
     }
 }
