+++ app/src/main/res/layout/fragment_delete_browsing_data.xml
@@ -29,7 +29,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical">
-            <org.mozilla.fenix.settings.DeleteBrowsingDataItem
+            <org.mozilla.fenix.settings.deletebrowsingdata.DeleteBrowsingDataItem
                 android:id="@+id/open_tabs_item"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -40,7 +40,7 @@
                 android:focusable="true"
                 app:deleteBrowsingDataItemTitle="@string/preferences_delete_browsing_data_tabs_title"
                 app:deleteBrowsingDataItemSubtitle="@string/preferences_delete_browsing_data_tabs_subtitle" />
-            <org.mozilla.fenix.settings.DeleteBrowsingDataItem
+            <org.mozilla.fenix.settings.deletebrowsingdata.DeleteBrowsingDataItem
                 android:id="@+id/browsing_data_item"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -51,7 +51,7 @@
                 android:focusable="true"
                 app:deleteBrowsingDataItemTitle="@string/preferences_delete_browsing_data_browsing_data_title"
                 app:deleteBrowsingDataItemSubtitle="@string/preferences_delete_browsing_data_browsing_data_subtitle" />
-            <org.mozilla.fenix.settings.DeleteBrowsingDataItem
+            <org.mozilla.fenix.settings.deletebrowsingdata.DeleteBrowsingDataItem
                 android:id="@+id/collections_item"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -62,7 +62,7 @@
                 android:focusable="true"
                 app:deleteBrowsingDataItemTitle="@string/preferences_delete_browsing_data_collections_title"
                 app:deleteBrowsingDataItemSubtitle="@string/preferences_delete_browsing_data_collections_subtitle" />
-            <org.mozilla.fenix.settings.DeleteBrowsingDataItem
+            <org.mozilla.fenix.settings.deletebrowsingdata.DeleteBrowsingDataItem
                 android:id="@+id/cookies_item"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -74,7 +74,7 @@
                 android:visibility="gone"
                 app:deleteBrowsingDataItemTitle="@string/preferences_delete_browsing_data_cookies"
                 app:deleteBrowsingDataItemSubtitle="@string/preferences_delete_browsing_data_cookies_subtitle" />
-            <org.mozilla.fenix.settings.DeleteBrowsingDataItem
+            <org.mozilla.fenix.settings.deletebrowsingdata.DeleteBrowsingDataItem
                 android:id="@+id/cached_files_item"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
@@ -86,7 +86,7 @@
                 android:visibility="gone"
                 app:deleteBrowsingDataItemTitle="@string/preferences_delete_browsing_data_cached_files"
                 app:deleteBrowsingDataItemSubtitle="@string/preferences_delete_browsing_data_cached_files_subtitle" />
-            <org.mozilla.fenix.settings.DeleteBrowsingDataItem
+            <org.mozilla.fenix.settings.deletebrowsingdata.DeleteBrowsingDataItem
                 android:id="@+id/site_permissions_item"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
