+++ app/src/test/java/org/mozilla/fenix/settings/account/AccountSettingsFragmentStoreTest.kt
@@ -2,16 +2,12 @@
    License, v. 2.0. If a copy of the MPL was not distributed with this
    file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.settings
+package org.mozilla.fenix.settings.account
 
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotSame
 import org.junit.Test
-import org.mozilla.fenix.settings.account.AccountSettingsFragmentAction
-import org.mozilla.fenix.settings.account.AccountSettingsFragmentState
-import org.mozilla.fenix.settings.account.AccountSettingsFragmentStore
-import org.mozilla.fenix.settings.account.LastSyncTime
 
 class AccountSettingsFragmentStoreTest {
