+++ app/src/test/java/org/mozilla/fenix/settings/account/AccountSettingsInteractorTest.kt
@@ -2,7 +2,7 @@
    License, v. 2.0. If a copy of the MPL was not distributed with this
    file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix.settings
+package org.mozilla.fenix.settings.account
 
 import androidx.navigation.NavController
 import androidx.navigation.NavDestination
@@ -12,10 +12,6 @@ import io.mockk.verify
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.mozilla.fenix.R
-import org.mozilla.fenix.settings.account.AccountSettingsFragmentAction
-import org.mozilla.fenix.settings.account.AccountSettingsFragmentDirections
-import org.mozilla.fenix.settings.account.AccountSettingsInteractor
-import org.mozilla.fenix.settings.account.AccountSettingsFragmentStore
 
 class AccountSettingsInteractorTest {
